WIP LteControlInfo: eliminate multicastGroupId, merge into destId
In LteControlInfo.msg, there is a multicastGroupId field, which if not -1,
indicates that the packet is targeted to multiple nodes in the RAN.
This is nonstandard -- real 5G NR appears to use special, dynamically
allocated node IDs (a MBMS RNTI) as dest address. Refactor the code,
and replace the multicastGroupId-based solution with dynamically allocated nodeIds.

Hints: When a new multicast IPv4 address is detected, the code (in IpToNic or Pdcp?)
currently computes a multicastGroupId. Replace this by allocating a
dest nodeId from Binder (Binder should manage the associations). Those new IDs
should be in a dedicated range, e.g. above 0x8000; add a utility function
in LteCommon.h to decide if a nodeId is a multicast addess. Then, everywhere
multicastGroupId is used, make the code test and use destId instead.