Pdcp: fix: fromDataPort() to use destId for MacCid
This is consistent with fromLowerLayer(), which uses sourceId.

To use ctrlInfoToMacCid() would be WRONG. It returns the UE's nodeId,
which ignores the fact that PDCP may have connections toward multiple
gNB/eNB entities due to Dual Connectivity and Handovers. (Handovers
are also offending becauses currently the old PDCP entities are NOT
deleted on handover!).

Note that for Dual Connectivity/Split Bearers, the split occurs (is
supposed to occur) in the PDCP layer, ie. PDCP may have to "legs",
ie. LCID is only valid in the context of the Master or Slave serving
entity.