MAJOR CHANGE: Introduce SMF-like functionality in Binder, and RRC modules in NIC
THIS IS A MAJOR ARCHITECTURAL CHANGE. Connections are not longer added
dynamically at each layer as packets arrive, but proactively as in
actual 5G systems. Right now it is triggered from the PDCP protocol of the node
where the packet enters the RAN, but that can be easily changed now into
e.g. static configuration, or dynamic configuration initiated by
some other entity.

Tricky parts of the implementation is:
- handling Dual Connectivity setups
- multicast
- D2D