IMPORTANT: macNodeId assignment moved to NED
macNodeIds are now generated and assigned to NED parameters using the
newly added simu5g_seq() NED function. It is no longer the Binder's
responsibility to assign nodeIds. The Binder now only checks that
the nodeIds that come from the NED parameters do not conflict and
correspond to the conventions (separate ranges for UEs, NodeBs, etc)

Motivation/Benefits:
- Removes initialization nightmare (see below)
- Allows manual nodeID assignment too (and mixed auto+manual too, in a
  somewhat limited way)

Original state:

The macNodeId parameter values set in NED or INI were IGNORED.
Ip2Nic invoked Binder's registerNode() function in INITSTAGE_LOCAL,
which generated a nodeId and the WROTE IT BACK into the module parameter
(that's why it was marked @mutable). Further modules in the NIC read that
parameter (IN THE SAME INIT STAGE) to obtain the nodeId (this made the
NIC sensitive to changes in the submodule order).