IMPORTANT: LteUe,NrNuE: ue.masterId --> ue.servingNodeId, nrServingNodeId
The parameter was renamed for clarity. In Simu5G, the word "master" was
used to mean both the Master eNodeB/gNodeB in a Dual Connectivity setup,
and the eNodeB/gNodeB to which the UE is connected. To reduce the chance
of confusion, the latter is now referred ti as "serving node" of the UE.