mimo: Partially delete incomplete MIMO implementation: muMimo NED parameter, TxMode constants, etc
MIMO is not supported in Simu5G, but a partial MIMO implementation is
present in the codebase, with a lot machinery that is NOT used in any
simulation (since they run with MIMO disabled). It is considered useful
both for users and for further development of Simu5G that this "crust"
is deleted from the codebase. It facilitates understanding of the code,
may speed up simulations (by removing computations whose result is then
never used by the simulation), etc.

This will be multiple commits -- this is the first one.

After the removal, simulation fingerprints remained unchanged, i.e.
everything works the same.

- Delete MIMO-related TxMode constants
- Remove dead conditional branches
- Delete muMimo parameter