LteFeedbackComputationRealistic: eliminate the use of random numbers for PMI
PMI (Precoding Matrix Indicator) values travel back to eNB in feedback
packets, but are NOT used as input for any computation there. Therefore,
we can set PMI to any value, it has no effect. This was verified by
changing intuniform(..) to e.g. 678+intuniform(...) -- fingerprints do
not change.

We elimimate the use of random numbers for PMI because (1) it is
unnecessary, and (2) it makes refactoring and potential future removal
of imcomplete MIMO support more difficult. Any change that adds/removes
drawing a random number affects further random numbers used in other
parts of the simulation, and thereby alters simulation trajectory. This
could be offset by using a different RNG stream here, but as said, using
RNG does not make sense here at all.